🚧 N0ZTF5 task: apply external agent result
