+++ .agentplane/tasks/202609090930-N0ZTF5/pr/diffstat.txt
@@ -1,3 +1,4 @@
+ .../commands/task/evidence-only-rework-commit.ts   | 147 +++++++++++++
  .../external-agent-implementation-recovery.test.ts | 228 +++++++++++++++++++++
- .../task/external-agent-implementation-recovery.ts | 160 ++++++++++++++-
- 2 files changed, 384 insertions(+), 4 deletions(-)
+ .../task/external-agent-implementation-recovery.ts |  29 ++-
+ 3 files changed, 400 insertions(+), 4 deletions(-)
