+++ .agentplane/tasks/202609090930-N0ZTF5/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202609090930-N0ZTF5/recover-committed-implementation-after-an-approv",
   "created_at": "2026-09-09T11:39:33.192Z",
-  "diffstat_sha256": "sha256:ceef6f8f3913be4cea3642f045b19784847a610f921206d24aa57d4731e524ee",
+  "diffstat_sha256": "sha256:203589408e7ee69295c36b94b34748e6ed25bcbb1cb20385727c5594d7b3de3e",
   "last_verified_at": "2026-09-09T13:33:35.796Z",
   "last_verified_diffstat_sha256": "sha256:ceef6f8f3913be4cea3642f045b19784847a610f921206d24aa57d4731e524ee",
   "pr_number": 5926,
