+++ .agentplane/tasks/202609090930-N0ZTF5/pr/review.md
@@ -29,9 +29,10 @@ Created: 2026-09-09T11:39:33.192Z
 - Head: computed live by `agentplane pr check` / `agentplane integrate`
 
 ```text
+ .../commands/task/evidence-only-rework-commit.ts   | 147 +++++++++++++
  .../external-agent-implementation-recovery.test.ts | 228 +++++++++++++++++++++
- .../task/external-agent-implementation-recovery.ts | 160 ++++++++++++++-
- 2 files changed, 384 insertions(+), 4 deletions(-)
+ .../task/external-agent-implementation-recovery.ts |  29 ++-
+ 3 files changed, 400 insertions(+), 4 deletions(-)
 ```
 
 </details>
